Added destructuring, Ramda + implementation.
