+++ package.json
@@ -33,8 +33,9 @@
   "dependencies": {
     "axios": "^0.17.1",
     "babel-loader": "^7.1.2",
-    "extract-text-webpack-plugin": "^3.0.2",
     "express": "^4.16.2",
+    "extract-text-webpack-plugin": "^3.0.2",
+    "ramda": "^0.25.0",
     "react": "^16.2.0",
     "react-dom": "^16.2.0",
     "webpack": "^3.10.0",
