+++ src/routes/services/account-services.js
@@ -6,21 +6,11 @@ const API_KEY = process.env.RIOT_API;
  * Performs service call to LoL API and obtains an account ID.
  * Requires a region name and summoner name.
  */
-getSummonerAccountInfo = function (region, name) {
-    const link = "https://" +  region + ".api.riotgames.com/lol/summoner/v3/summoners/by-name/" + name;
-    return new Promise((resolve, reject) => {
-        axios({
-            method: "GET",
-            url: link,
-            params: {'api_key' : API_KEY}
-        }).then( result => {
-            resolve(result.data.accountId);
-        }).catch( err => {
-            reject(err); 
-        });
-    });
+module.exports = { 
+    async getSummonerAccountInfo (region, name) {
+        const link = "https://" +  region + ".api.riotgames.com/lol/summoner/v3/summoners/by-name/" + name;
+        const {data : {accountId}} =  await axios.get(link, {params: { 'api_key' : API_KEY }});
+        return accountId; 
+    }
 }
 
-module.exports = {
-    getSummonerAccountInfo : getSummonerAccountInfo
-}
