+++ src/routes/services/static-services.js
@@ -7,21 +7,10 @@ const API_KEY = process.env.RIOT_API;
  * Performs service call to lol api and gets a champions name based on ID.
  * Requires a region name and summoner name.
  */
-getChampionName = function (region, id) {
-    const link = "https://" + region + ".api.riotgames.com/lol/static-data/v3/champions/" + id + "?locale=en_US";
-    return new Promise((resolve, reject) => {
-        axios({
-            method: "GET",
-            url: link,
-            params: {'api_key' : API_KEY}
-        }).then( result => {
-            resolve(result.data.name);
-        }).catch( err => {
-            reject(err); 
-        });
-    });
-}
-
-module.exports = {
-    getChampionName : getChampionName
+module.exports = { 
+    async getChampionName (region, id) {
+        const link = "https://" + region + ".api.riotgames.com/lol/static-data/v3/champions/" + id + "?locale=en_US";
+        const {data : {name}} =  await axios.get(link, {params: { 'api_key' : API_KEY }});
+        return name;
+    }
 }
