+++ webpack.config.js
@@ -23,7 +23,7 @@ module.exports = {
     module: { 
         rules: [ 
             {
-                test: /\.js$/,
+                test: /\.jsx?$/,
                 exclude: /(node_modules|bower_components)/,
                 use: {
                     loader: 'babel-loader',
